*: implement an async mutable plugin interface
All plugin run_* methods now have mutable access to their state. This
enables use-cases such as in-memory caches.

A symptom of using an asynchronous API throughout the the whole
processing chain is that the configured plugin collection must be
statically initialized, so that the references are guaranteed to be
valid for the whole lifetime of the program.

Changeset
* Implement the async interface and enhance its tests
* Port the channel-filter
* Port the quay metadata plugin
* Port edge_add_remove
* Port node_remove to async
* Port external web plugin example
* Adapt the graph-builder:
  It was necessary to move the AppSettings to the GB library and
  initialized statically as it is accessed by the static initialisation
  of the configured plugins
* Adapt the policy-engine